moved around some of the visualization files
