+++ src/training/training_utils.py
@@ -242,7 +242,7 @@ def create_experiment_dir(args):
     now = datetime.now()
     dt_string = now.strftime("%d-%m-%Y__%H-%M-%S")
 
-    experiment_dir = dt_string + "_experiment_" + experiment_id
+    experiment_dir = "experiments/" + dt_string + "_experiment_" + experiment_id
 
     os.makedirs(experiment_dir)
     os.makedirs(os.path.join(experiment_dir, "checkpoints/"), exist_ok=True)
