+++ src/visualization/plot_gradcam_amp_n_std.py
@@ -18,9 +18,11 @@ def generate_gradcam_amp_plot(df, experiment_string, save_dir):
 
 if __name__ == "__main__":
     csv_filename = (
-        r"./../training/test_experiment_cifar10_cutout_resnet18/cifar10_cutout_resnet18_b.csv"
+        # r"./../training/test_experiment_cifar10_cutout_resnet18/cifar10_cutout_resnet18_b.csv"
+        r"05-12-2020__07-57-58_experiment_cifar10_cutout_resnet18\cifar10_cutout_resnet18_a.csv"
+
     )
-    experiment_string = "model_b"
+    experiment_string = "model_a"
     save_dir = os.path.dirname(csv_filename)
 
     df = load_csv_into_dataframe(csv_filename)
